LayoutTests/media/media-fragments/TC0087-expected.txt



Title: TC0087
Fragment: 't=3,7&foo'
Comment: Rubbish after & is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.75') OK
END OF TEST
